Spring Framework 5 - Beginner ToGuru/recipe-reactive-mongodb/src/main/java/com/fredrikpedersen/services/UnitOfMeasureServiceImpl.java
package com.fredrikpedersen.services;

import com.fredrikpedersen.converters.UnitOfMeasureToUnitOfMeasureCommand;
import com.fredrikpedersen.commands.UnitOfMeasureCommand;
import com.fredrikpedersen.repositories.UnitOfMeasureRepository;
import com.fredrikpedersen.repositories.reactive.UnitOfMeasureReactiveRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import reactor.core.publisher.Flux;

import java.util.Set;
import java.util.stream.Collectors;
import java.util.stream.StreamSupport;

@Service
@RequiredArgsConstructor
public class UnitOfMeasureServiceImpl implements UnitOfMeasureService {

    private final UnitOfMeasureReactiveRepository unitOfMeasureRepository;
    private final UnitOfMeasureToUnitOfMeasureCommand unitOfMeasureToUnitOfMeasureCommand;

    @Override
    public Flux<UnitOfMeasureCommand> listAllUoms() {

        return unitOfMeasureRepository
                .findAll()
                .map(unitOfMeasureToUnitOfMeasureCommand::convert);
    }
}
