Spring Framework 5 - Beginner ToGuru/recipe-reactive-mongodb/src/test/java/com/fredrikpedersen/services/UnitOfMeasureServiceImplTest.java
package com.fredrikpedersen.services;

import com.fredrikpedersen.commands.UnitOfMeasureCommand;
import com.fredrikpedersen.converters.UnitOfMeasureToUnitOfMeasureCommand;
import com.fredrikpedersen.domain.UnitOfMeasure;
import com.fredrikpedersen.repositories.reactive.UnitOfMeasureReactiveRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import reactor.core.publisher.Flux;

import java.util.List;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

/**
 * @author Fredrik Pedersen
 * @version 1.0
 * @since 14/07/2021 at 10:01
 */

class UnitOfMeasureServiceImplTest {

    private final UnitOfMeasureToUnitOfMeasureCommand unitOfMeasureToUnitOfMeasureCommand = new UnitOfMeasureToUnitOfMeasureCommand();
    private UnitOfMeasureService service;

    @Mock
    private UnitOfMeasureReactiveRepository unitOfMeasureReactiveRepository;

    @BeforeEach
    public void setUp() {
        MockitoAnnotations.openMocks(this);
        service = new UnitOfMeasureServiceImpl(unitOfMeasureReactiveRepository, unitOfMeasureToUnitOfMeasureCommand);
    }

    @Test
    public void listAllUoms() {

        //given
        final UnitOfMeasure uom1 = UnitOfMeasure.builder().id("1").build();
        final UnitOfMeasure uom2 = UnitOfMeasure.builder().id("1").build();

        when(unitOfMeasureReactiveRepository.findAll()).thenReturn(Flux.just(uom1, uom2));

        //when
        final List<UnitOfMeasureCommand> commands = service.listAllUoms().collectList().block();

        //then
        assertEquals(2, commands.size());
        verify(unitOfMeasureReactiveRepository, times(1)).findAll();
    }

}